create a seperate base class + make S3Fileio pluggable
